mengubah semua metode delete

+++ src/component/email/MailingListMemberList.js
@@ -9,6 +9,8 @@ import AuthenticationService from "./../../logic/AuthenticationService";
 function MailingListMemberList({ state, dispatch }) {
   const [data, setData] = useState([]);
   const [loading, setLoading] = useState(true);
+  const [isDelete, setIsDelete] = useState(false);
+  const [isDeleteData, setIsDeleteData] = useState()
   const mailingListId = state.mailingListId;
   const axiosConfig = AuthenticationService.getAxiosConfig();
   //console.log(mailingListId);
@@ -23,8 +25,7 @@ function MailingListMemberList({ state, dispatch }) {
           params: { mailinglist_id: mailingListId },
         }
       );
-      //console.log(res.data);
-      console.log(`useEffect axios ${mailingListId}`);
+      console.log(res, "check list mailinglistmember")
 
       if (res.status === 200) {
         setData(res.data);
@@ -41,21 +42,54 @@ function MailingListMemberList({ state, dispatch }) {
     });
   }, []);
   const deleteData = async (data) => {
-    try {
-      const res = await axios.delete(`${config.SERVER_URL}mailinglistmember`, {
-        ...axiosConfig,
-        params: { mailinglist_id: mailingListId, email: data.email },
-      });
-      console.log(res.data);
-      setData(res.data);
-    } catch (e) {
-      console.log(e);
-    }
+    setIsDeleteData(data)
+    setIsDelete(true)
+    // try {
+    //   const res = await axios.delete(`${config.SERVER_URL}mailinglistmember`, {
+    //     ...axiosConfig,
+    //     params: { mailinglist_id: mailingListId, email: data.email },
+    //   });
+      
+    //   setData(res.data);
+    // } catch (e) {
+    //   console.log(e);
+    // }
   };
+  const handleDelete = async (e) => {
+    if(e.currentTarget.textContent.toUpperCase() === "YES"){
+      try {
+        const res = await axios.delete(`${config.SERVER_URL}mailinglistmember`, {
+          ...axiosConfig,
+          params: { mailinglist_id: mailingListId, email: isDeleteData.email },
+        });
+        setIsDelete(false)
+        setData(res.data);
+      } catch (e) {
+        console.log(e);
+      }
+    }else{
+      setIsDelete(false)
+    }
+  }
   return (
     <React.Fragment>
       {!loading ? (
-        <section className="content">
+        <section className="content"style={{position:"relative"}}>
+          <div style={{position:"absolute", zIndex: "10", backgroundColor: "rgb(0,0,0, 0.5)", height: "100%", width:"100rem", display: isDelete ? "block": "none"}} role="dialog">
+          <div className={`${isDelete ? "" : "modal"} position-absolute`}style={{position:"absolute", zIndex: "11", top:"50%", transform: "translateY(-50%)", left:"0", right: "0", margin: "auto"}} tabindex="-1" role="dialog">
+              <div className="modal-dialog " role="document">
+                <div className="modal-content">
+                  <div className="modal-body">
+                    <p>Are you sure you wish to delete this item?</p>
+                  </div>
+                  <div className="modal-footer">
+                    <button type="button" className="btn btn-primary" onClick={handleDelete}>Yes</button>
+                    <button type="button" className="btn btn-secondary" data-dismiss="modal" onClick={handleDelete}>No</button>
+                  </div>
+                </div>
+              </div>
+            </div>
+          </div>
           <div className="container-fluid">
             <div className="row clearfix">
               <div className="col-lg-12 col-md-12 col-sm-12 col-xs-12">
@@ -82,12 +116,7 @@ function MailingListMemberList({ state, dispatch }) {
                                 <button
                                   className="btn btn-danger waves-effect "
                                   onClick={() => {
-                                    if (
-                                      window.confirm(
-                                        "Are you sure you wish to delete this item?"
-                                      )
-                                    )
-                                      deleteData(i);
+                                    deleteData(i);
                                   }}
                                 >
                                   Delete
